Refactor Pipeline out of Store
Fixes #30

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/debrief/rap-lines/issues/30?shareId=XXXX-XXXX-XXXX-XXXX).
